feat(trailing-comma): use correct trailing comma rule (#6)
BREAKING CHANGE: This rule may introduce some breaking changes
since the versioning of tslint-config-airbnb does not adhere to the
semantic version rules. So a minor update of tslint config could
introduce breaking changes. Since there was an update of tslint config,
this creates a new major. The new rule does not break a build,
and is newly added to tslint-config-airbnb.